@@ -1,3 +1,3 @@
 module github.com/verystar/nacos-go-sdk
 
-go 1.15
+go 1.17
